+++ CHANGES.md
@@ -1,1 +1,11 @@
 ### Changelogs
+
+##### 2.4.2_1
++ Workaround for issue #1 by adding to `application.conf`:
+
+  ```
+  play.modules.disabled += "play.api.db.evolutions.EvolutionsModule"
+  ```
+
+##### 2.4.2
+initial release